+++ src/commanders/mod.rs
@@ -1,6 +1,5 @@
 mod _cmd;
 
-pub mod chapterer;
 pub mod fps_adjuster;
 pub mod fps_changer;
 pub mod fps_reader;